requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyMySQL==1.1.0
python-decouple==3.8
Pillow>=10.0.0
django-cors-headers==4.3.1
cryptography==41.0.7
django-filter==23.5
channels==4.0.0
daphne==4.0.0
social-auth-app-django==5.4.0
azure-ai-inference
PyPDF2
requests
python-dotenv
google-generativeai
google-api-python-client
google-auth
google-auth-oauthlib
